VolumeInfo: Rewrite test cases to use XML files
The tool VolumeInfo was rewritten to be able to generate an XML file with the output of the Win32 APIs. This allows the VolumeInfo tool to be used to capture data which is helpful later in creating and then running unit tests on a developer machine without access to the Operating System for testing (e.g. WinXP). This makes the tests more reliable and trustworthy as there is less room for error when copying the results from one machine to another (prior by eyeballs and fingers, now using the API results themselves to write the XML). It can also help in debugging issues in the future should they occur on remote machines the developers don't have direct access to.

To enable this option, use the syntax:

  VolumeInfo -l <path>

The output will be written to a file called `VolumeInfoTest.<machinename>.xml`.

Test cases for Windows XP SP3 and Windows 10 v2004 were replaced with outputs from the XML file when run. More test cases were easier to add.

Issue: DOTNET-206