VolumeInfo: Support devices that don't provide volume mount information
The ImDisk driver for a RAMDISK doesn't support getting volume information. MSDN states that this must be supported, so is an error in the driver itself. But see if we can support some basic functionality none-the-less, as it's usable and some information is obtainable.

Quote from https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/mountmgr/ni-mountmgr-ioctl_mountdev_query_device_name

  Support for this IOCTL by the mount manager clients is mandatory. Upon
  receiving this IOCTL a client driver must provide the (nonpersistent)
  device (or target) name for the volume. The mount manager uses the device
  name returned by the client as the target of a symbolic link. An example of
  a device name would be "\Device\HarddiskVolume1".

Issue: DOTNET-212