VolumeInfo: Reorganize VolumeDeviceInfo properties into groups
Instead of having all properties "flat" in the class, put them behind further classes/properties to try and group the values. This makes collecting the data more intuitive.

The PartitionInformation class was moved into the Win32 folder, as it should no longer be needed for classes, except for providing their own implementtion of IOSVolumeInformation.

Issue: DOTNET-217