VolumeInfo: Support Floppy disk drives
Floppy disks don't return partition information, so the Partition property is now set to null, so it can be easily tested prior (without having to use an exception). The main program now checks this, so it shouldn't crash immediately.

Floppy disks don't return device query information. So obtain the removable drive media implicitly by the fact that the geometry says it's a floppy disk (or it's marked there as removable). In case that the geometry is also not available (i.e. a disk is not inserted), then we have to do some heuristics.

Floppy disks don't reliably return if media is present or not (it always returns false), so we have to use file attributes to get this information, and if the root volume is a directory. This is how DriveInfo from .NET does the check. Verified and it works on a USB Floppy too.

Added test cases for Windows XP (tested with an emulated floppy).

Issue: DOTNET-221