VolumeInfo: Suppress dialog popups when detecting floppy disks
In the commit a8f78997 ("VolumeInfo: Support Floppy disk drives"), we need to check the attributes if the drive is present, because the other IOCTLs don't work. On Windows XP, this has the side effect of showing a pop-up window, which should be avoided. Indeeded, in MSDN sources, they make note of this.

https://referencesource.microsoft.com/#mscorlib/system/io/file.cs,56cd161c65ab07fe

    // For floppy drives, normally the OS will pop up a dialog saying
    // there is no disk in drive A:, please insert one.  We don't want that.
    // SetErrorMode will let us disable this, but we should set the error
    // mode back, since this may have wide-ranging effects.
    int oldMode = Win32Native.SetErrorMode(Win32Native.SEM_FAILCRITICALERRORS);

Issue: DOTNET-221