VolumeInfo: Set the version to 1.0.0-alpha.1
At this time, the program is still pretty much work in progress, while researching a few things, and there are still some bugs, so it's definitely in "Alpha" stage. By providing a version, we can use the same label in the VCS, which makes it easier to upload binaries.

Issue: DOTNET-230