VolumeInfo: Mark Win32 as internal
There are a lot of internal functionality, but still which needs to be tested by providing simulations. There's no reasonable OpenSource mock tool found that can do this, so it was all made previously public.

Having to use the InternalsVisibleTo attribute seems like a weird reverse dependency and should be avoided, but the benefits are pretty high here. So make the Win32 namespace internal, and make those internals visible to the unit test cases for providing simulations.

Issue: DOTNET-233