VolumeInfo: Get the disk extents
The disk extents for a basic disk are usually the same information as the partition. But Windows supports volumes that span multiple disks (dynamic disks), and so there can be more than one disk for a volume. Provide a mechanism to get all the disk extents, which can be used to determine the device used for the underlying storage.

Issue: DOTNET-225