VolumeInfo: Reorganize Device information inside Disk.Device property
The disk device information often isn't available for floppy drives and multi-partition and multi-disk volumes. Move this to its own property and set it to null if this information isn't available.
* Vendor ID, Product ID, Product Revision, Serial Number, Scsi Type/Modifier, BusType, Command Queueing
* Device Type, Device Number, Device GUID (Partition Number isn't used here, but use the partition structure instead)

The BytsPerPhysicalSector also isn't available in some cases, but we assume then the same as the logical sector.

If the user really wants this information, they should use the Extents and get the information from the \\.\PhysicalDriveX path.

Issue: DOTNET-226