VolumeInfo: Move Win32 specific functionality into a new class
Make the VolumeDeviceInfo class public abstract, that allows other developers to inherit this for their own unit testing, where they might not have access to direct hardware, or running on a CI machine. It also allows easier implementation later for other flavours of operating systems as can be detected.

There is now a new static Create() method which should be used to instantiate the object. This supports easy updates in the future by returning a Task<> object, instead of having a blocking constructor.