VolumeInfo: Remove the DeviceType and DeviceNumber
The DeviceType contained information already similar to the DriveType. The DeviceNumber is unique for the driver, not the actual DeviceType as suggested by MSDN documentation (so a non-Microsoft driver can return that it is also a Disk and return the Device Number 0, just exactly like the ImDisk RAM Disk Driver, and there's a conflict).

Therefore, this information is not useful and it's best that it's removed.

Issue: DOTNET-234