VolumeInfo: Provide test results with Win10 v2004 and BTRFS
We see that the logical bytes per sector is 4096, when it should perhaps be 512. However, I consider this a bug in the Btrfs implementation that it is different to the underlying physical volume. We can't change it, as this would result in an incorrect calculation of the disk size (Cyl * Trk/Cyl * Sec/Trk * Bytes/Sec) which is consistent (relatively) to the rest of the size calculations.

Issue: DOTNET-237