VolumeDeviceInfo: If SignAssembly is set, don't use InternalsVisibleTo
The `InternalsVisibleTo` only works if the public key token is provided when strong signing (`/p:SignAssembly=true /p:AssemblyOriginatorKeyFile=xxx.snk`). So if the user provides this in the project or on the command line, set the constant SIGN, so that this can be used to disable test cases and code that rely on internal details.

Issue: DOTNET-427