VolumeDeviceInfo: Add ByPassIO for Windows 11 to the FileSystemFlags
Windows 11 introduced a new flag on NTFS. Add this flag to the enumeration that it can be reported.

Running VolumeInfo on the file system would just report a number, not the features, so manual parsing of the enumeration is now done for the case of new flags added in the future.

Issue: DOTNET-460