VolumeDeviceInfo: SYSLIB004 ReliabilityContract is only for .NET Framework
.NET Core doesn't support Constrained Execution Regions.

The overly complicated code for .NET Framework is also not needed, as the `ReleaseHandle` is already running in a Constrained Execution Region. So no need for the `try`/`finally` code section. Also the checks of the `IntPtr` isn't needed, as the `SafeHandle` will never call `ReleaseHandle` after it's freed.

Issue: DOTNET-937